// src/interfaces/reqRes.ts
//Hay servicios que permiten tipar las apis de forma automática que hacerla manualmente
//-->https://app.quicktype.io/<---

/* export interface ReqResListado {
  page: number;
  per_page: number;
  total: number;
  total_pages: number;
  data: Datum[];
  support: Support;
}

export interface Datum {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  avatar: string;
}

export interface Support {
  url: string;
  text: string;
}
 */
// Generated by https://quicktype.io --> Se realiza por la extensión de VSC con CTRL+SHIFT+V+ENTER (Teniendo previamente copiado la información)

// Generated by https://quicktype.io

export interface TopLevel {
  page: number;
  per_page: number;
  total: number;
  total_pages: number;
  data: User[];
  support: Support;
}

export interface User {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  avatar: string;
}

export interface Support {
  url: string;
  text: string;
}